requirements.txt
django
python-decouple
psycopg2-binary
Pillow
twilio
django-allauth